optimize the regcpy for TO_REF; move csr reg write to exe stage; raise interrupt for ref

+++ tools/spike-diff/difftest.cc
@@ -94,7 +94,7 @@ void sim_t::diff_get_regs(void* diff_context) {
 void sim_t::diff_set_regs(void* diff_context) {
   struct diff_context_t* ctx = (struct diff_context_t*)diff_context;
   for (int i = 0; i < NR_GPR; i++) {
-    state->XPR.write(i, (sreg_t)ctx->gpr[i]);
+    state->XPR.write(i, (sreg_t)((int32_t)ctx->gpr[i]));
   }
   state->pc = ctx->pc;
   CSR_LIST(CSR_SET)
